src/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
configparser==5.0.0
Django==2.2.4
django-cors-headers==2.4.0
django-extensions==2.1.6
django-filter==2.1.0
django-minio-storage==0.3.7
django-model-utils==3.1.2
django-oauth-toolkit==1.3.2
djangorestframework==3.9.1
djangorestframework-camel-case==1.0.3
djangorestframework-simplejwt==3.3
drf-nested-routers==0.91
gunicorn==19.9.0
idna==2.9
minio==5.0.10
oauthlib==3.1.0
psycopg2-binary==2.7.5
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
